move Bitcast, SExt, ZExt and Trunc to shared file, fix broken reductions, add to tryGetConstantSignedInteger
