+++ jlm/llvm/ir/TraceTests.cpp
@@ -5,6 +5,7 @@
 
 #include <gtest/gtest.h>
 
+#include <jlm/llvm/ir/operators/Bitcast.hpp>
 #include <jlm/llvm/ir/operators/IntegerOperations.hpp>
 #include <jlm/llvm/ir/operators/IOBarrier.hpp>
 #include <jlm/llvm/ir/operators/lambda.hpp>
@@ -132,4 +133,72 @@ TEST(TraceTests, testGetConstantSignedInteger)
   // A match output is not a constant integer, neither is the lambda output
   EXPECT_EQ(tryGetConstantSignedInteger(*matchOutput), std::nullopt);
   EXPECT_EQ(tryGetConstantSignedInteger(*lambdaOutput), std::nullopt);
+}
+
+TEST(TraceTests, testGetConstantSignedIntegerExtAndTrunc)
+{
+  using namespace jlm;
+  using namespace jlm::llvm;
+
+  /**
+   * Creates an RVSDG graph that look like:
+   *
+   * c = BITS32(5)
+   * sext = SExt(32 -> 64) c  // Should be 5
+   * zext = ZExt(32 -> 64) c  // Should be 5
+   *
+   * c2 = BITS8(-20)          // Should be -20  (0b1110 1100)
+   * sext2 = SExt(8 -> 32) c2 // Should be -20  (0b1111 .... 1110 1100)
+   * zext2 = ZExt(8 -> 32) c2 // Should be 236  (0b0000 .... 1110 1100)
+   *
+   * c3 = BITS32(1023)
+   * trunc3 = Trunc(32 -> 8)       // Should be -1   (0b1111 1111)
+   * sext3 = SExt(8 -> 32) trunc3  // Should be -1   (0b1111 .... 1111 1111)
+   * zext3 = ZExt(8 -> 32) trunc3  // Should be  255 (0b0000 .... 1111 1111)
+   *
+   * and uses tryGetConstantSignedInteger to get the integer values of the different constants.
+   */
+
+  // Arrange
+  rvsdg::Graph graph;
+
+  const auto int64Type = rvsdg::BitType::Create(64);
+  const auto int32Type = rvsdg::BitType::Create(32);
+  const auto int8Type = rvsdg::BitType::Create(8);
+
+  auto & bits32Output5 = rvsdg::BitConstantOperation::create(
+      graph.GetRootRegion(),
+      rvsdg::BitValueRepresentation(32, 5));
+  auto & sextOutput = SExtOperation::create(64, bits32Output5);
+  auto & zextOutput = ZExtOperation::create(64, bits32Output5);
+
+  auto & bits8OutputMinus20 = rvsdg::BitConstantOperation::create(
+      graph.GetRootRegion(),
+      rvsdg::BitValueRepresentation(8, -20));
+  auto & sext2Output = SExtOperation::create(32, bits8OutputMinus20);
+  auto & zext2Output = ZExtOperation::create(32, bits8OutputMinus20);
+
+  auto & bits32Output1023 = rvsdg::BitConstantOperation::create(
+      graph.GetRootRegion(),
+      rvsdg::BitValueRepresentation(32, 1023));
+  auto & truncOutput = TruncOperation::create(8, bits32Output1023);
+  auto & sext3Output = SExtOperation::create(32, truncOutput);
+  auto & zext3Output = ZExtOperation::create(32, truncOutput);
+
+  // Assert
+  // c = BITS32(5), sext = SExt(32 -> 64), zext = ZExt(32 -> 64)
+  EXPECT_EQ(tryGetConstantSignedInteger(bits32Output5), 5);
+  EXPECT_EQ(tryGetConstantSignedInteger(sextOutput), 5);
+  EXPECT_EQ(tryGetConstantSignedInteger(zextOutput), 5);
+
+  // c2 = BITS8(-20), sext2 = SExt(8 -> 32), zext2 = ZExt(8 -> 32)
+  EXPECT_EQ(tryGetConstantSignedInteger(bits8OutputMinus20), -20);
+  EXPECT_EQ(tryGetConstantSignedInteger(sext2Output), -20);
+  EXPECT_EQ(tryGetConstantSignedInteger(zext2Output), 236);
+
+  // c3 = BITS32(1023), trunc3 = Trunc(32 -> 8), sext3 = SExt(8 -> 32), zext3 = ZExt(8 -> 32)
+  EXPECT_EQ(tryGetConstantSignedInteger(bits32Output1023), 1023);
+  EXPECT_EQ(tryGetConstantSignedInteger(truncOutput), -1);
+  EXPECT_EQ(tryGetConstantSignedInteger(sext3Output), -1);
+  EXPECT_EQ(tryGetConstantSignedInteger(zext3Output), 255);
 }
