+++ jlm/llvm/ir/operators/operators.hpp
@@ -7,6 +7,7 @@
 #define JLM_LLVM_IR_OPERATORS_OPERATORS_HPP
 
 #include <jlm/llvm/ir/ipgraph-module.hpp>
+#include <jlm/llvm/ir/operators/Bitcast.hpp>
 #include <jlm/llvm/ir/tac.hpp>
 #include <jlm/llvm/ir/types.hpp>
 #include <jlm/rvsdg/binary.hpp>
@@ -772,109 +773,6 @@ private:
   ICmpPredicate predicate_;
 };
 
-class ZExtOperation final : public rvsdg::UnaryOperation
-{
-public:
-  ~ZExtOperation() noexcept override;
-
-  ZExtOperation(size_t nsrcbits, size_t ndstbits)
-      : UnaryOperation(rvsdg::BitType::Create(nsrcbits), rvsdg::BitType::Create(ndstbits))
-  {
-    if (ndstbits < nsrcbits)
-      throw util::Error("# destination bits must be greater than # source bits.");
-  }
-
-  ZExtOperation(
-      const std::shared_ptr<const jlm::rvsdg::BitType> & srctype,
-      const std::shared_ptr<const jlm::rvsdg::BitType> & dsttype)
-      : UnaryOperation(srctype, dsttype)
-  {
-    if (dsttype->nbits() < srctype->nbits())
-      throw util::Error("# destination bits must be greater than # source bits.");
-  }
-
-  ZExtOperation(
-      std::shared_ptr<const jlm::rvsdg::Type> srctype,
-      std::shared_ptr<const jlm::rvsdg::Type> dsttype)
-      : UnaryOperation(srctype, dsttype)
-  {
-    auto st = dynamic_cast<const jlm::rvsdg::BitType *>(srctype.get());
-    if (!st)
-      throw util::Error("expected bitstring type.");
-
-    auto dt = dynamic_cast<const jlm::rvsdg::BitType *>(dsttype.get());
-    if (!dt)
-      throw util::Error("expected bitstring type.");
-
-    if (dt->nbits() < st->nbits())
-      throw util::Error("# destination bits must be greater than # source bits.");
-  }
-
-  bool
-  operator==(const Operation & other) const noexcept override;
-
-  [[nodiscard]] std::string
-  debug_string() const override;
-
-  [[nodiscard]] std::unique_ptr<Operation>
-  copy() const override;
-
-  jlm::rvsdg::unop_reduction_path_t
-  can_reduce_operand(const jlm::rvsdg::Output * operand) const noexcept override;
-
-  jlm::rvsdg::Output *
-  reduce_operand(jlm::rvsdg::unop_reduction_path_t path, jlm::rvsdg::Output * operand)
-      const override;
-
-  inline size_t
-  nsrcbits() const noexcept
-  {
-    return std::static_pointer_cast<const rvsdg::BitType>(argument(0))->nbits();
-  }
-
-  inline size_t
-  ndstbits() const noexcept
-  {
-    return std::static_pointer_cast<const rvsdg::BitType>(result(0))->nbits();
-  }
-
-  static std::unique_ptr<llvm::ThreeAddressCode>
-  create(const Variable * operand, const std::shared_ptr<const jlm::rvsdg::Type> & type)
-  {
-    auto operandBitType = CheckAndExtractBitType(operand->Type());
-    auto resultBitType = CheckAndExtractBitType(type);
-
-    auto operation =
-        std::make_unique<ZExtOperation>(std::move(operandBitType), std::move(resultBitType));
-    return ThreeAddressCode::create(std::move(operation), { operand });
-  }
-
-  static rvsdg::Output &
-  Create(rvsdg::Output & operand, const std::shared_ptr<const rvsdg::Type> & resultType)
-  {
-    auto operandBitType = CheckAndExtractBitType(operand.Type());
-    auto resultBitType = CheckAndExtractBitType(resultType);
-
-    return *rvsdg::CreateOpNode<ZExtOperation>(
-                { &operand },
-                std::move(operandBitType),
-                std::move(resultBitType))
-                .output(0);
-  }
-
-private:
-  static std::shared_ptr<const rvsdg::BitType>
-  CheckAndExtractBitType(const std::shared_ptr<const rvsdg::Type> & type)
-  {
-    if (auto bitType = std::dynamic_pointer_cast<const rvsdg::BitType>(type))
-    {
-      return bitType;
-    }
-
-    throw util::TypeError("BitType", type->debug_string());
-  }
-};
-
 /* floating point constant operator */
 
 class ConstantFP final : public rvsdg::SimpleOperation
@@ -1543,77 +1441,6 @@ public:
   }
 };
 
-class BitCastOperation final : public rvsdg::UnaryOperation
-{
-public:
-  ~BitCastOperation() noexcept override;
-
-  BitCastOperation(
-      std::shared_ptr<const jlm::rvsdg::Type> srctype,
-      std::shared_ptr<const jlm::rvsdg::Type> dsttype)
-      : UnaryOperation(srctype, dsttype)
-  {
-    check_types(srctype, dsttype);
-  }
-
-  BitCastOperation(const BitCastOperation &) = default;
-
-  explicit BitCastOperation(Operation &&) = delete;
-
-  BitCastOperation &
-  operator=(const Operation &) = delete;
-
-  BitCastOperation &
-  operator=(Operation &&) = delete;
-
-  bool
-  operator==(const Operation & other) const noexcept override;
-
-  [[nodiscard]] std::string
-  debug_string() const override;
-
-  [[nodiscard]] std::unique_ptr<Operation>
-  copy() const override;
-
-  jlm::rvsdg::unop_reduction_path_t
-  can_reduce_operand(const jlm::rvsdg::Output * output) const noexcept override;
-
-  jlm::rvsdg::Output *
-  reduce_operand(jlm::rvsdg::unop_reduction_path_t path, jlm::rvsdg::Output * output)
-      const override;
-
-  static std::unique_ptr<llvm::ThreeAddressCode>
-  create(const Variable * operand, std::shared_ptr<const jlm::rvsdg::Type> type)
-  {
-    auto pair = check_types(operand->Type(), type);
-
-    auto op = std::make_unique<BitCastOperation>(pair.first, pair.second);
-    return ThreeAddressCode::create(std::move(op), { operand });
-  }
-
-  static jlm::rvsdg::Output *
-  create(jlm::rvsdg::Output * operand, std::shared_ptr<const jlm::rvsdg::Type> rtype)
-  {
-    auto pair = check_types(operand->Type(), rtype);
-    return rvsdg::CreateOpNode<BitCastOperation>({ operand }, pair.first, pair.second).output(0);
-  }
-
-private:
-  static std::pair<std::shared_ptr<const jlm::rvsdg::Type>, std::shared_ptr<const jlm::rvsdg::Type>>
-  check_types(
-      const std::shared_ptr<const jlm::rvsdg::Type> & otype,
-      const std::shared_ptr<const jlm::rvsdg::Type> & rtype)
-  {
-    if (otype->Kind() != rvsdg::TypeKind::Value)
-      throw util::Error("expected value type.");
-
-    if (rtype->Kind() != rvsdg::TypeKind::Value)
-      throw util::Error("expected value type.");
-
-    return std::make_pair(otype, rtype);
-  }
-};
-
 class ConstantStruct final : public rvsdg::SimpleOperation
 {
 public:
@@ -1679,95 +1506,6 @@ private:
     }
 
     throw util::TypeError("StructType", type->debug_string());
-  }
-};
-
-class TruncOperation final : public rvsdg::UnaryOperation
-{
-public:
-  ~TruncOperation() noexcept override;
-
-  TruncOperation(
-      const std::shared_ptr<const jlm::rvsdg::BitType> & otype,
-      const std::shared_ptr<const jlm::rvsdg::BitType> & rtype)
-      : UnaryOperation(otype, rtype)
-  {
-    if (otype->nbits() < rtype->nbits())
-      throw util::Error("expected operand's #bits to be larger than results' #bits.");
-  }
-
-  TruncOperation(
-      std::shared_ptr<const jlm::rvsdg::Type> optype,
-      std::shared_ptr<const jlm::rvsdg::Type> restype)
-      : UnaryOperation(optype, restype)
-  {
-    auto ot = dynamic_cast<const jlm::rvsdg::BitType *>(optype.get());
-    if (!ot)
-      throw util::Error("expected bits type.");
-
-    auto rt = dynamic_cast<const jlm::rvsdg::BitType *>(restype.get());
-    if (!rt)
-      throw util::Error("expected bits type.");
-
-    if (ot->nbits() < rt->nbits())
-      throw util::Error("expected operand's #bits to be larger than results' #bits.");
-  }
-
-  bool
-  operator==(const Operation & other) const noexcept override;
-
-  [[nodiscard]] std::string
-  debug_string() const override;
-
-  [[nodiscard]] std::unique_ptr<Operation>
-  copy() const override;
-
-  jlm::rvsdg::unop_reduction_path_t
-  can_reduce_operand(const jlm::rvsdg::Output * operand) const noexcept override;
-
-  jlm::rvsdg::Output *
-  reduce_operand(jlm::rvsdg::unop_reduction_path_t path, jlm::rvsdg::Output * operand)
-      const override;
-
-  inline size_t
-  nsrcbits() const noexcept
-  {
-    return std::static_pointer_cast<const rvsdg::BitType>(argument(0))->nbits();
-  }
-
-  inline size_t
-  ndstbits() const noexcept
-  {
-    return std::static_pointer_cast<const rvsdg::BitType>(result(0))->nbits();
-  }
-
-  static std::unique_ptr<llvm::ThreeAddressCode>
-  create(const Variable * operand, const std::shared_ptr<const jlm::rvsdg::Type> & type)
-  {
-    auto ot = std::dynamic_pointer_cast<const jlm::rvsdg::BitType>(operand->Type());
-    if (!ot)
-      throw util::Error("expected bits type.");
-
-    auto rt = std::dynamic_pointer_cast<const jlm::rvsdg::BitType>(type);
-    if (!rt)
-      throw util::Error("expected bits type.");
-
-    auto op = std::make_unique<TruncOperation>(std::move(ot), std::move(rt));
-    return ThreeAddressCode::create(std::move(op), { operand });
-  }
-
-  static jlm::rvsdg::Output *
-  create(size_t ndstbits, jlm::rvsdg::Output * operand)
-  {
-    auto ot = std::dynamic_pointer_cast<const jlm::rvsdg::BitType>(operand->Type());
-    if (!ot)
-      throw util::Error("expected bits type.");
-
-    return rvsdg::CreateOpNode<TruncOperation>(
-               { operand },
-               std::move(ot),
-               rvsdg::BitType::Create(ndstbits))
-        .output(0);
   }
 };
